requirements.txt
fastapi 
uvicorn
pypacker
IPy
requests
sqlalchemy 
passlib[bcrypt] 
pyjwt